+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/actions/list/cds-action-replace-bot-v2/cds-action-replace-bot-v2.component.ts
@@ -7,6 +7,7 @@ import { LoggerInstance } from 'src/chat21-core/providers/logger/loggerInstance'
 import { FaqService } from 'src/app/services/faq.service';
 import { Faq } from 'src/app/models/faq-model';
 import { Intent } from 'src/app/models/intent-model';
+import { MatCheckbox } from '@angular/material/checkbox';
 
 @Component({
   selector: 'cds-action-replace-bot-v2',
@@ -20,16 +21,16 @@ export class CdsActionReplaceBotV2Component implements OnInit, OnChanges {
   @Output() updateAndSaveAction = new EventEmitter();
 
   //bots: Chatbot[] = [];
-  chatbots_name_list: Array<{name: string, value: string, id: string, icon?:string}>;
+  chatbots_name_list: Array<{name: string, value: string, id: string, slug: string, disabled: boolean, icon?:string}>;
   chatbots_block_name_list: Array<{name: string, value: string, icon?:string}>;
-  bot_selected: Chatbot;
+  bot_selected: {name: string, value: string, id: string, slug: string, disabled: boolean, icon?:string}
 
   private logger: LoggerService = LoggerInstance.getInstance();
   
   constructor(
     private chatbotService: FaqKbService,
     private faqService: FaqService
-    ) { }
+  ) { }
 
   ngOnInit(): void {
     this.logger.log("[ACTION REPLACE BOT] action (on-init): ", this.action)
@@ -43,9 +44,14 @@ export class CdsActionReplaceBotV2Component implements OnInit, OnChanges {
   async initialize(){
     await this.getAllBots();
     if(this.action && this.action.botName){
-      let selectedChatbot = this.chatbots_name_list.find(el => el.name === this.action.botName)
-      if(selectedChatbot){
-        this.getAllFaqById(selectedChatbot.id)
+      if(this.action.nameAsSlug){
+        this.bot_selected = this.chatbots_name_list.find(el => el.slug === this.action.botName)
+      }
+      if(!this.action.nameAsSlug){
+        this.bot_selected = this.chatbots_name_list.find(el => el.name === this.action.botName)
+      }
+      if(this.bot_selected){
+        this.getAllFaqById(this.bot_selected.id)
       }
     }
   }
@@ -56,7 +62,7 @@ export class CdsActionReplaceBotV2Component implements OnInit, OnChanges {
       this.chatbotService.getFaqKbByProjectId().subscribe({ next: (chatbots) => {
         this.logger.log("[ACTION REPLACE BOT] chatbots: ", chatbots);
         //this.bots = bots;
-        this.chatbots_name_list = chatbots.map(a => ({ name: a.name, value: a.name, id: a._id, icon: 'smart_toy'}));
+        this.chatbots_name_list = chatbots.map(a => ({ name: a.name, value: a.name, slug: a.slug, id: a._id, disabled: this.action.nameAsSlug? !a.slug: false, icon: 'smart_toy'}));
         resolve(true)
       }, error: (error) => {
         this.logger.error("[ACTION REPLACE BOT] error get bots: ", error);
@@ -81,8 +87,12 @@ export class CdsActionReplaceBotV2Component implements OnInit, OnChanges {
     }})
   }
 
-  onChangeSelect(event: {name: string, value: string, id: string}) {
+  onChangeSelect(event: {name: string, value: string, slug: string, id: string}) {
     this.logger.log("[ACTION REPLACE BOT] onChangeActionButton event: ", event)
+    this.bot_selected = this.chatbots_name_list.find(el => el.id === event.id)
+    if(this.action.nameAsSlug){
+      this.action.botName = event.slug;
+    }
     this.action.botName = event.value;
     this.getAllFaqById(event.id)
     this.updateAndSaveAction.emit()
@@ -95,9 +105,40 @@ export class CdsActionReplaceBotV2Component implements OnInit, OnChanges {
     this.logger.log("[ACTION REPLACE BOT] action selet block to execute: ", this.action)
   }
 
-  onResetBlockSelect(event){
-    this.action.blockName = null;
+  onResetSelect(event, key: string){
+    switch(key){
+      case 'botName':
+        this.action.botName = null;
+        this.bot_selected = null;
+        break;
+      case 'blockName':
+        this.action.blockName = null
+        break;   
+    }
     this.updateAndSaveAction.emit()
+  }
+
+  onChangeCheckbox(event: MatCheckbox, target){
+    this.action[target] = !this.action[target];
+    this.chatbots_name_list = this.chatbots_name_list.map(a => ({ name: a.name, value: a.name, slug: a.slug, id: a.id, disabled: this.action.nameAsSlug? !a.slug: false, icon: 'smart_toy'}));
+    if(event.checked && this.action.botName){
+      this.action.botName = this.getChatbotByNameOrSlug()
+    }
+    this.updateAndSaveAction.emit()
+  }
+
+
+  private getChatbotByNameOrSlug(): string{
+    let chatbotByName = this.chatbots_name_list.find(el => el.name === this.action.botName)
+    if(chatbotByName){
+      this.bot_selected = chatbotByName
+      return chatbotByName.slug
+    }
+    let chatbotBySlug = this.chatbots_name_list.find(el => el.slug === this.action.botName)
+    if(chatbotBySlug){
+      this.bot_selected = chatbotBySlug
+      return chatbotBySlug.slug
+    }
   }