+++ src/app/models/action-model.ts
@@ -236,6 +236,7 @@ export class ActionReplaceBot extends Action {
 export class ActionReplaceBotV2 extends Action {
     botName: string;
     blockName: string;
+    nameAsSlug: boolean
     constructor(){
         super();
         this._tdActionType = TYPE_ACTION.REPLACE_BOTV2;